Include details and actions config in refactor
